+++ libft/src/ft_putaddr.c
@@ -6,7 +6,7 @@
 /*   By: tdefresn <tdefresn@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/01/14 00:40:45 by tdefresn          #+#    #+#             */
-/*   Updated: 2016/01/14 04:49:21 by tdefresn         ###   ########.fr       */
+/*   Updated: 2016/01/14 15:24:59 by tdefresn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	ft_putaddr(const void *addr)
 {
 	char	*str;
 
-	str = ft_itoa_base((int)addr, 16);
+	str = ft_itoa_base((uint64_t)addr, 16);
 	ft_putstr("0x");
 	ft_putstr(str);
 }
